fix: fixing random errors and renaming folderrs
